+++ deploy/cttso-ica-to-pieriandx-cdk/lambdas/layers/lambda_utils/globals.py
@@ -43,6 +43,7 @@ PIERIANDX_LAMBDA_LAUNCH_FUNCTION_ARN_SSM_PATH = "cttso-ica-to-pieriandx-lambda-f
 
 MAX_ATTEMPTS_GET_CASES = 5
 LIST_CASES_RETRY_TIME = 5
+MAX_SUBMISSIONS = 3
 
 LOGGER_STYLE = "%(asctime)s - %(levelname)-8s - %(module)-25s - %(funcName)-40s : LineNo. %(lineno)-4d - %(message)s"
 # Redcap lambda path