+++ deploy/cttso-ica-to-pieriandx-cdk/Changelog.md
@@ -12,6 +12,7 @@ are nested under deploy/cttso-ica-to-pieriandx-cdk.
 
 * Reduce number of items returned by portal from 1000 to 100 to prevent internal server errors.
 * Version control gspread to prevent pip install failure for lambdas
+* Reduce number of submissions to pieriandx to reduce timeouts 
 
 ## 2022-09-22